new naming + documentation for AgentGraph/GraphAgent
